[fix] Setting 폴더 viewcontroller의 Extension, Layout 분류 및 폴더 재분류
